variable for lang_card_min_height, move theme to language chooser
